Split out `chakraComponents` into separate files
