+++ src/use-chakra-select-props.ts
@@ -27,9 +27,11 @@ const useChakraSelectProps = <
   onBlur,
   ...props
 }: Props<Option, IsMulti, Group>): Props<Option, IsMulti, Group> => {
-  // Combine the props passed into the component with the props that can be set
-  // on a surrounding form control to get the values of `isDisabled` and
-  // `isInvalid`
+  /**
+   * Combine the props passed into the component with the props that can be set
+   * on a surrounding form control to get the values of `isDisabled` and
+   * `isInvalid`
+   */
   const inputProps = useFormControl({
     id: inputId,
     isDisabled,
@@ -45,15 +47,17 @@ const useChakraSelectProps = <
    */
   const placeholderColor = useColorModeValue("gray.400", "whiteAlpha.400");
 
-  // Ensure that the size used is one of the options, either `sm`, `md`, or `lg`
+  /** Ensure that the size used is one of the options, either `sm`, `md`, or `lg` */
   let realSize: Size = size;
   const sizeOptions: Size[] = ["sm", "md", "lg"];
   if (!sizeOptions.includes(size)) {
     realSize = "md";
   }
 
-  // Ensure that the tag variant used is one of the options, either `subtle`,
-  // `solid`, or `outline` (or undefined)
+  /**
+   * Ensure that the tag variant used is one of the options, either `subtle`,
+   * `solid`, or `outline` (or undefined)
+   */
   let realTagVariant: TagVariant | undefined = tagVariant;
   const tagVariantOptions: TagVariant[] = ["subtle", "solid", "outline"];
   if (tagVariant !== undefined) {
@@ -62,14 +66,14 @@ const useChakraSelectProps = <
     }
   }
 
-  // Ensure that the selected option style is either `color` or `check`
+  /** Ensure that the selected option style is either `color` or `check` */
   let realSelectedOptionStyle: SelectedOptionStyle = selectedOptionStyle;
   const selectedOptionStyleOptions: SelectedOptionStyle[] = ["color", "check"];
   if (!selectedOptionStyleOptions.includes(selectedOptionStyle)) {
     realSelectedOptionStyle = "color";
   }
 
-  // Ensure that the color used for the selected options is a string
+  /** Ensure that the color used for the selected options is a string */
   let realSelectedOptionColor: string = selectedOptionColor;
   if (typeof selectedOptionColor !== "string") {
     realSelectedOptionColor = "blue";
